Makueni NCD dataset added
This study seeks to find out the prevalence of hypertension and diabetes in Makueni County and visualize the prevalence of diabetes and hypertension in Makueni County. The study found that DM and HTN in Makueni County are most prevalent in Makueni sub-county and least prevalent in Kibwezi East sub-county. It was also established that cases of DM and HTN are more prevalent among females compared to their male counterparts. Compared to men, women are more likely to be diagnosed with a combination of hypertension and diabetes. It is also observed that there is direct correlation between the frequency of clinic visits and the cases of DM and HTN.